fix entities to render cycles
