+++ packages/ketcher-core/src/domain/helpers/__tests__/rna.test.ts
@@ -1,6 +1,6 @@
 import { getPeptideLibraryItem, getRnaPartLibraryItem } from '../rna';
 import { CoreEditor } from 'application/editor/internal';
-import { KetMonomerClass } from 'application/formatters/types/ket';
+import { KetMonomerClass } from 'domain/constants/monomers';
 
 describe('RNA/Peptide Helper Functions', () => {
   let mockEditor: CoreEditor;
